+++ Summoner's Companion/Converters/ArrayToStringConverter.cs
@@ -11,8 +11,9 @@ namespace Summoner_s_Companion.Converters
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
             var array = ((IEnumerable) value).OfType<object>();
+            var seperator = parameter == null ? ", " : (string)parameter;
             if (!array.Any()) return "";
-            return array.Aggregate((x, y) => x + ", " + y);
+            return array.Aggregate((x, y) => x + seperator + y);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)