+++ Summoner's Companion/Converters/ItemImageCOnverter.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Globalization;
+using System.Windows.Data;
+
+namespace Summoner_s_Companion.Converters
+{
+    public class ItemImageConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return $"http://ddragon.leagueoflegends.com/cdn/6.24.1/img/item/{value}.png";
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}