+++ Summoner's Companion/Models/RelayCommand.cs
@@ -5,39 +5,29 @@ namespace Summoner_s_Companion.Models
 {
     public class RelayCommand : ICommand
     {
+        private readonly Action<object> _execute;
+        private readonly Func<object, bool> _canExecute;
+
         public event EventHandler CanExecuteChanged
         {
             add => CommandManager.RequerySuggested += value;
             remove => CommandManager.RequerySuggested -= value;
         }
 
-        private readonly Action _methodToExecute;
-        private readonly Func<bool> _canExecuteEvaluator;
-
-        public RelayCommand(Action methodToExecute, Func<bool> canExecuteEvaluator)
+        public RelayCommand(Action<object> execute, Func<object, bool> canExecute = null)
         {
-            _methodToExecute = methodToExecute;
-            _canExecuteEvaluator = canExecuteEvaluator;
-        }
-
-        public RelayCommand(Action methodToExecute)
-            : this(methodToExecute, null)
-        {
+            _execute = execute;
+            _canExecute = canExecute;
         }
 
         public bool CanExecute(object parameter)
         {
-            if (_canExecuteEvaluator == null)
-            {
-                return true;
-            }
-            bool result = _canExecuteEvaluator.Invoke();
-            return result;
+            return _canExecute == null || _canExecute(parameter);
         }
 
         public void Execute(object parameter)
         {
-            _methodToExecute.Invoke();
+            _execute(parameter);
         }
     }
 }